fix: add review suggestions
